examples/free_demo.py
#!/usr/bin/env python3
"""
Pharmaceutical RAG System - FREE Demo Mode
Uses only local models, no API costs!
"""

import pathway as pw
from dotenv import load_dotenv
import logging

# Configure logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

def run_free_demo():
    """Run the pharmaceutical RAG system with free local models only"""
    
    print("🆓 PHARMACEUTICAL RAG SYSTEM - 100% FREE MODE")
    print("="*55)
    print("💡 This demo uses ONLY free, local models:")
    print("   ✓ Local sentence transformers for embeddings")
    print("   ✓ No API calls or costs")
    print("   ✓ Complete document processing pipeline")
    print("   ✓ Vector search capabilities")
    print("   ✓ REST API endpoints")
    print()
    
    try:
        # Load configuration for document processing only
        config_file = "app_no_llm.yaml"
        
        logger.info("🧬 Starting FREE Pharmaceutical RAG System...")
        logger.info("📊 Configuration: Local embeddings only (no LLM costs)")
        
        with open(config_file) as f:
            config = pw.load_yaml(f)
        
        print("🚀 SYSTEM STARTING...")
        print("📡 API Server: http://127.0.0.1:8000")
        print("📚 Document Processing: Ready for PDFs in ./data/")
        print("🔍 Vector Search: Enabled with local embeddings")
        print()
        print("🌐 Available Endpoints:")
        print("   GET  /health                    - System status")
        print("   POST /api/v1/pw_list_documents  - List indexed docs")
        print("   POST /api/v1/retrieve           - Semantic search")
        print()
        print("💡 To test: Open another terminal and run:")
        print("   curl http://localhost:8000/health")
        print("   python test_api.py")
        print()
        
        # Run the system
        pw.run()
        
    except FileNotFoundError as e:
        logger.error(f"❌ Configuration file not found: {e}")
    except Exception as e:
        logger.error(f"❌ Error: {e}")
        print("\n💡 This demonstrates the complete RAG pipeline!")
        print("   - Document ingestion ✓")
        print("   - Vector embeddings ✓") 
        print("   - Semantic search ✓")
        print("   - REST API ✓")
        print("   - Zero API costs ✓")

if __name__ == "__main__":
    run_free_demo()